feat: Modulo de paltos creado exitosamente

+++ src/dish/infraestructure/repositories/dish.mongo.repository.ts
@@ -24,7 +24,7 @@ export class DishMongoRepository implements IDishRepository {
             if (existingDish){
                 this.logger.warn(`Intento de crear plato con nombre duplicado: ${dish.name}`);
                 throw new Error('DUPLICATE_ENTRY');
-            }
+            };
 
             const newDish = new this.dishModel({
                 ...dish,
@@ -33,11 +33,8 @@ export class DishMongoRepository implements IDishRepository {
             });
 
             await newDish.save();
-
             this.logger.log(`Plato creado exitosamente con ID: ${id} y nombre: ${dish.name}`);
-
             return newDish as IDish;
-
         } catch (error) {
             if (error.message === 'DUPLICATE_ENTRY') throw error;
             
@@ -50,8 +47,8 @@ export class DishMongoRepository implements IDishRepository {
         try {
             const dishes = await this.dishModel.find();
 
+            this.logger.log('Platos obtenidos exitosamente');
             return dishes as IDish[];
-
         } catch (error) {
             this.logger.error('Error técnico al obtener los platos', error?.stack);
             throw new Error('DATABASE_ERROR');
@@ -65,10 +62,9 @@ export class DishMongoRepository implements IDishRepository {
                 this.logger.warn(`Plato con ID ${id} no encontrado`);
                 return null;
             };
-            this.logger.log(`Plato encontrado con ID: ${id}`);
 
+            this.logger.log(`Plato encontrado con ID: ${id}`);
              return dish as IDish;
-          
         } catch (error) {
             this.logger.error(`Error al acceder a la base de datos: ${error.message}`);
             throw new Error('DATABASE_ERROR');
@@ -78,7 +74,6 @@ export class DishMongoRepository implements IDishRepository {
     async update(id: string, dish: PatchDishInput): Promise<IDishPatch | null> {
         try {
            const existingDish = await this.dishModel.findOne({ id });
-
            if (!existingDish) {
                 this.logger.warn(`Intento de actualizar plato inexistente con ID: ${id}`);
                 return null;
@@ -93,7 +88,6 @@ export class DishMongoRepository implements IDishRepository {
            this.logger.log(`Plato actualizado exitosamente con ID: ${id}`);
 
            return updateDish as IDishPatch;
-
         } catch (error) {
             this.logger.error(`Error al actualizar el plato: ${id}`, error?.stack);
             throw new Error('DATABASE_ERROR');
@@ -102,12 +96,15 @@ export class DishMongoRepository implements IDishRepository {
 
     async delete(id: string): Promise<boolean>{
         try {
-            const result = await this.dishModel.deleteOne({ id }).exec();
+            const existingDish = await this.dishModel.findOne({ id });
+            if (!existingDish) {
+                this.logger.warn(`Intento de eliminar plato inexistente con ID: ${id}`);
+                return false;
+            };
 
+            await this.dishModel.deleteOne({ id }).exec();
             this.logger.log(`Plato eliminado exitosamente con ID: ${id}`);
-            
-            return result.deletedCount > 0;
-            
+            return true;
         }catch (error){
             this.logger.error(`Error al eliminar el plato: ${id}`, error?.stack);
             throw new Error('DATABASE_ERROR');
